clean up fine tune example for PR
