+++ .env.example
@@ -11,8 +11,10 @@ CDP_API_KEY_PRIVATE_KEY=your_cdp_api_key_private_key
 # Hyperbolic (Required)
 HYPERBOLIC_API_KEY=your_hyperbolic_api_key
 
+# RSA key file path (Required)
 SSH_PRIVATE_KEY_PATH=~/.ssh/id_rsa
 
+# LangChain (Required for Finetune tool)
 LANGCHAIN_TRACING_V2=true
 LANGCHAIN_ENDPOINT="your_langchain_endpoint"
 LANGCHAIN_API_KEY="your_langchain_api_key"
@@ -50,7 +52,7 @@ USE_RETWEET=true
 USE_DEPLOY_MULTITOKEN=true
 USE_WEB_SEARCH=true
 USE_REQUEST_TOOLS=true
-USE_FINETUNE_TOOLS=true
+USE_FINETUNE_TOOLS=false
 
 
 ELEVEN_API_KEY=your_eleven_api_key
