+++ hyperbolic_agentkit_core/actions/__init__.py
@@ -8,6 +8,7 @@ from hyperbolic_agentkit_core.actions.get_spend_history import GetSpendHistoryAc
 from hyperbolic_agentkit_core.actions.terminate_compute import TerminateComputeAction
 from hyperbolic_agentkit_core.actions.get_current_balance import GetCurrentBalanceAction
 from hyperbolic_agentkit_core.actions.link_wallet_address import LinkWalletAddressAction
+from hyperbolic_agentkit_core.actions.remote_finetune_example import RunFinetuneAction
 # WARNING: All new HyperbolicAction subclasses must be imported above, otherwise they will not be discovered
 # by get_all_hyperbolic_actions(). The import ensures the class is registered as a subclass of HyperbolicAction.
 def get_all_hyperbolic_actions() -> list[type[HyperbolicAction]]:
@@ -22,5 +23,5 @@ HYPERBOLIC_ACTIONS = get_all_hyperbolic_actions()
 
 __all__ = [
     "HYPERBOLIC_ACTIONS", "HyperbolicAction", "RentComputeAction", "GetAvailableGpusAction",
-    "GetGpuStatusAction", "SSHAccessAction", "RemoteShellAction", "TerminateComputeAction", "GetSpendHistoryAction", "GetCurrentBalanceAction", "LinkWalletAddressAction"
+    "GetGpuStatusAction", "SSHAccessAction", "RemoteShellAction", "TerminateComputeAction", "GetSpendHistoryAction", "GetCurrentBalanceAction", "LinkWalletAddressAction", "RunFinetuneAction"
 ]
